requirements.txt
playwright>=1.42.0
pytest>=8.1.1
pytest-playwright>=0.4.4
pytest-xdist>=3.5.0
allure-pytest>=2.13.5
